requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
imbalanced-learn
mlflow
pyyaml
pytest